Move types to types/explore
